Update SimpleBanking.c
Note: This project will be drastically improved.
Added error checking in the program so that, for e.g. , the user cannot withdraw money if the balance is below 0. Also, some parts of the code were changed because they were slightly incorrect and this commit mainly consists of slight improvements and error checking code.