Update SimpleBanking.c
Added an option for the user to pay a bill. Error checking was done so that the user cannot pay the bill if the account balance will be less than 0. Also, when the user pays the bill, the total amount will be put inside an array that holds all of the previous withdraws