Update LottoGame.c
Error checking done so that if the user enters the same number or the user enters a number greater than 42 or less than 0, the array that stores the entered numbers from the user will not be filled with these incorrect numbers. So if the user enters e.g. 1,2,2,3,4,5, this array will not be saved. The frequency of these numbers will not be saved. Then, if the user enters 1,2,3,4,5,6, the numbers will then be saved and the frequency will be saved.