Update SimpleBanking.c
Fixed the gets() error I got when asking the user to enter a message. This was because I had to use white space characters because the user can enter a sentence, which contains white spaces. Then, I added a getchar() before the gets() so that the input could be gathered and so that the buffer doesn't jump away from this line.