Update CurrencyConversion.c
Added comments to the program. Also, fixed the average calculation so that it doesn't divide by 3.0 (this was because the program has been extended to allow the user to enter up to 50 values, when before, the user had to enter only 3). Also, removed "int i;" declaration and put it inside the for loop so it looks cleaner.